Rename error folder to exception
